Use watcher binary
