+++ package.json
@@ -8,9 +8,9 @@
     "build": "npm-run-all --parallel compile:* && npm-run-all --parallel build:*",
     "build:client": "node ./scripts/build.js",
     "build:server": "tsc -p ./src/tsconfig.json",
-    "watch": "NODE_ENV=development ts-node ./scripts/watch.ts",
+    "watch": "NODE_ENV=development ts-node ./scripts/watcher/bin/watcher.ts ./config/watcher.ts",
     "compile:css-types": "tcm src/core/client/",
-    "compile:relay-stream": "relay-compiler --src ./src/core/client/stream --schema $(ts-node -T ./scripts/schemaPath.ts tenant) --language typescript --artifactDirectory ./src/core/client/stream/__generated__ --no-watchman",
+    "compile:relay-stream": "relay-compiler --src ./src/core/client/stream --schema $(ts-node ./scripts/schemaPath.ts tenant) --language typescript --artifactDirectory ./src/core/client/stream/__generated__ --no-watchman",
     "start:development": "NODE_ENV=development ts-node --project ./src/tsconfig.json -r tsconfig-paths/register ./src/index.ts",
     "start:webpackDevServer": "node ./scripts/start.js",
     "lint-fix": "npm run lint:server -- --fix && npm run lint:client -- --fix && npm run lint:scripts -- --fix",
