--- scripts/watch.ts
@@ -1,4 +1,0 @@
-import config from "../config/watcher";
-import watch from "./watcher";
-
-watch(config);
