+++ scripts/watcher/index.ts
@@ -2,4 +2,4 @@ export { default as ChokidarWatcher } from "./ChokidarWatcher";
 export { default as CommandExecutor } from "./CommandExecutor";
 export { default as RestartingExecutor } from "./RestartingExecutor";
 export { default } from "./watch";
-export * from "./watch";
+export * from "./types";
